# app/Http/Controllers/Api/AuthController.php
<?php

namespace App\Http\Controllers\Api;

use App\Helpers\ApiResponse;
use App\Helpers\FcmNotificationHelper;
use App\Http\Controllers\Controller;
use App\Models\Setting;
use App\Models\User;
use Carbon\Carbon;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\DB;
use Illuminate\Support\Facades\Hash;
use Illuminate\Support\Facades\Log;
use Illuminate\Support\Facades\Mail;
use Illuminate\Support\Facades\Validator;
use Illuminate\Support\Str;

class AuthController extends Controller
{
    /**
     * @OA\Post(
     *     path="/api/v1/login",
     *     tags={"Auth"},
     *     summary="Login user",
     *     description="Authenticate user and return Sanctum token.",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"email","password"},
     *             @OA\Property(property="email", type="string", format="email", example="john@example.com"),
     *             @OA\Property(property="password", type="string", format="password", example="secret123"),
     *             @OA\Property(property="fcm_token", type="string", nullable=true, example="fcm_token_value"),
     *             @OA\Property(property="device_id", type="string", nullable=true, example="device-001"),
     *             @OA\Property(property="platform", type="string", nullable=true, example="android")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Login successful"
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Invalid credentials or validation error"
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error"
     *     )
     * )
     */
    public function login(Request $request)
    {
        try {

            $validated = Validator::make($request->all(), [
                'email' => ['required', 'email'],
                'password' => ['required'],
                'fcm_token' => ['nullable', 'string'],
                'device_id' => ['nullable', 'string', 'max:255'],
                'platform' => ['nullable', 'string', 'max:50'],
            ]);

            if ($validated->fails()) {
                return ApiResponse::error('Validation Error', $validated->errors(), 401);
            }

            $user = User::where('email', $request->email)->first();

            if (! $user || ! Hash::check($request->password, $user->password)) {
                return ApiResponse::error('Invalid credentials', null, 401);
            }

            $token = $user->createToken('flutter-app')->plainTextToken;

            return ApiResponse::success([
                'user' => $user,
                'permissions' => $user->getPermissionsViaRoles(),
                'token' => $token,
            ], 'Login successful');
        } catch (\Exception $e) {
            return ApiResponse::error($e->getMessage(), null, 500);
        }
    }

    /**
     * @OA\Get(
     *     path="/api/v1/me",
     *     tags={"Auth"},
     *     summary="Get authenticated user",
     *     security={{"sanctum":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="User information retrieved successfully"
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthenticated"
     *     )
     * )
     */
    public function me(Request $request)
    {
        $user = $request->user();
        // FcmNotificationHelper::storeTokenForUser(
        //     $user,
        //     $request->input('fcm_token'),
        //     $request->input('device_id'),
        //     $request->input('platform')
        // );

        // FcmNotificationHelper::sendToLoggedInUser(
        //     'New Update',
        //     'You have a new notification. Thank you for login.',
        //     ['type' => 'general']
        // );

        return ApiResponse::success([
            'user' => $user,
        ], 'User information retrieved successfully');
    }

    /**
     * @OA\Post(
     *     path="/api/v1/fcm-token",
     *     tags={"Auth"},
     *     summary="Store FCM token for authenticated user",
     *     security={{"sanctum":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"fcm_token"},
     *             @OA\Property(property="fcm_token", type="string", example="fcm_token_value"),
     *             @OA\Property(property="device_id", type="string", nullable=true, example="device-001"),
     *             @OA\Property(property="platform", type="string", nullable=true, example="android")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="FCM token stored successfully"
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthenticated"
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error"
     *     )
     * )
     */
    public function storeFcmToken(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'fcm_token' => ['required', 'string'],
            'device_id' => ['nullable', 'string', 'max:255'],
            'platform' => ['nullable', 'string', 'max:50'],
        ]);

        if ($validator->fails()) {
            return ApiResponse::error('Validation Error', $validator->errors(), 422);
        }

        $user = $request->user();
        if (! $user) {
            return ApiResponse::error('Unauthenticated.', null, 401);
        }

        $tokenRecord = FcmNotificationHelper::storeTokenForUser(
            $user,
            $request->input('fcm_token'),
            $request->input('device_id'),
            $request->input('platform')
        );

        if (! $tokenRecord) {
            return ApiResponse::error('Failed to store FCM token.', null, 500);
        }

        return ApiResponse::success([
            'id' => $tokenRecord->id,
            'user_id' => $tokenRecord->user_id,
            'token' => $tokenRecord->token,
            'device_id' => $tokenRecord->device_id,
            'platform' => $tokenRecord->platform,
            'is_active' => (bool) $tokenRecord->is_active,
            'last_used_at' => optional($tokenRecord->last_used_at)?->toISOString(),
            'created_at' => optional($tokenRecord->created_at)?->toISOString(),
            'updated_at' => optional($tokenRecord->updated_at)?->toISOString(),
        ], 'FCM token stored successfully.');
    }

    /**
     * @OA\Post(
     *     path="/api/v1/logout",
     *     tags={"Auth"},
     *     summary="Logout current device",
     *     security={{"sanctum":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Logged out successfully"
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthenticated"
     *     )
     * )
     */
    public function logout(Request $request)
    {
        $request->user()->currentAccessToken()->delete();

        return ApiResponse::success(null, 'Logged out successfully');
    }

    /**
     * @OA\Post(
     *     path="/api/v1/logout-all",
     *     tags={"Auth"},
     *     summary="Logout all devices",
     *     security={{"sanctum":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Logged out from all devices successfully"
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthenticated"
     *     )
     * )
     */
    public function logoutAll(Request $request)
    {
        $request->user()->tokens()->delete();

        return ApiResponse::success(null, 'Logged out from all devices successfully');
    }

    /**
     * Handle forgot password request
     *
     * @OA\Post(
     *     path="/api/v1/forgot-password",
     *     tags={"Auth"},
     *     summary="Send password reset link",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"email"},
     *             @OA\Property(property="email", type="string", format="email", example="john@example.com")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Password reset link sent"
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error"
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error"
     *     )
     * )
     */
    public function forgotPassword(Request $request)
    {
        try {
            // Validate the request
            $validated = Validator::make($request->all(), [
                'email' => 'required|email|exists:users,email',
            ], [
                'email.required' => 'The email field is required.',
                'email.email' => 'Please enter a valid email address.',
                'email.exists' => 'We could not find a user with that email address.',
            ]);

            if ($validated->fails()) {
                return ApiResponse::error('Validation Error', $validated->errors(), 422);
            }

            // Generate a password reset token
            $token = Str::random(64);

            // Store the token in the password_resets table
            DB::table('password_resets')->updateOrInsert(
                ['email' => $request->email],
                [
                    'email' => $request->email,
                    'token' => Hash::make($token),
                    'created_at' => Carbon::now(),
                ]
            );

            // Send the password reset email
            $user = User::where('email', $request->email)->first();
            $resetUrl = url("/api/v1/reset-password?token={$token}&email={$request->email}");
            $companyName = Setting::get('company_name', 'Technofra Admin');

            try {
                Mail::send('emails.password-reset', [
                    'user' => $user,
                    'resetUrl' => $resetUrl,
                    'token' => $token,
                    'companyName' => $companyName,
                ], function ($message) use ($user, $companyName) {
                    $message->to($user->email, $user->name)
                        ->subject('Reset Your Password - ' . $companyName);
                });

                return ApiResponse::success(null, 'We have sent a password reset link to your email address.');
            } catch (\Exception $e) {
                // Log the error for debugging
                Log::error('Password reset email failed: ' . $e->getMessage());

                return ApiResponse::error('Failed to send password reset email. Please try again.', null, 500);
            }
        } catch (\Exception $e) {
            return ApiResponse::error($e->getMessage(), null, 500);
        }
    }

    /**
     * Handle password reset
     *
     * @OA\Post(
     *     path="/api/v1/reset-password",
     *     tags={"Auth"},
     *     summary="Reset password",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"token","email","password","password_confirmation"},
     *             @OA\Property(property="token", type="string", example="generated_reset_token"),
     *             @OA\Property(property="email", type="string", format="email", example="john@example.com"),
     *             @OA\Property(property="password", type="string", format="password", example="new-password"),
     *             @OA\Property(property="password_confirmation", type="string", format="password", example="new-password")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Password reset successful"
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Invalid token or validation error"
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error"
     *     )
     * )
     */
    public function resetPassword(Request $request)
    {
        try {
            // Validate the request
            $validated = Validator::make($request->all(), [
                'token' => 'required',
                'email' => 'required|email|exists:users,email',
                'password' => 'required|string|min:6|confirmed',
            ], [
                'email.required' => 'The email field is required.',
                'email.email' => 'Please enter a valid email address.',
                'email.exists' => 'We could not find a user with that email address.',
                'password.required' => 'The password field is required.',
                'password.min' => 'Password must be at least 6 characters.',
                'password.confirmed' => 'Password confirmation does not match.',
            ]);

            if ($validated->fails()) {
                return ApiResponse::error('Validation Error', $validated->errors(), 422);
            }

            // Verify the token
            $passwordReset = DB::table('password_resets')
                ->where('email', $request->email)
                ->first();

            if (! $passwordReset || ! Hash::check($request->token, $passwordReset->token)) {
                return ApiResponse::error('Invalid or expired password reset token.', null, 422);
            }

            // Check if token is expired (24 hours)
            if (Carbon::parse($passwordReset->created_at)->addHours(24)->isPast()) {
                DB::table('password_resets')->where('email', $request->email)->delete();

                return ApiResponse::error('Password reset token has expired. Please request a new one.', null, 422);
            }

            // Update the user's password
            $user = User::where('email', $request->email)->first();
            $user->password = Hash::make($request->password);
            $user->save();

            // Delete the password reset token
            DB::table('password_resets')->where('email', $request->email)->delete();

            return ApiResponse::success(null, 'Your password has been reset successfully. You can now login with your new password.');
        } catch (\Exception $e) {
            return ApiResponse::error($e->getMessage(), null, 500);
        }
    }
}